+++ tests/Providers/CouchbaseProvider.php
@@ -10,36 +10,67 @@ class CouchbaseProvider extends AdapterProvider
     public function __construct()
     {
         if (class_exists('CouchbaseCluster')) {
+            // Couchbase SDK <=2.2
             $cluster = new \CouchbaseCluster('couchbase://couchbase:11210?detailed_errcodes=1');
+            $authenticator = new \Couchbase\PasswordAuthenticator();
+            $authenticator->username('Administrator')->password('password');
+            $cluster->authenticate($authenticator);
+            $bucket = $cluster->openBucket('default');
+            $collection = $bucket;
+            $bucketManager = $bucket->manager();
+        } elseif (class_exists('Couchbase\ClusterOptions')) {
+            // Couchbase SDK >=3.0
+            $options = new \Couchbase\ClusterOptions();
+            $options->credentials('Administrator', 'password');
+            $cluster = new \Couchbase\Cluster('couchbase://couchbase:11210?detailed_errcodes=1', $options);
+            $bucket = $cluster->bucket('default');
+            $collection = $bucket->defaultCollection();
+            $bucketManager = $cluster->buckets();
         } elseif (class_exists('\Couchbase\Cluster')) {
+            // Couchbase SDK >=2.3 & <3.0
             $cluster = new \Couchbase\Cluster('couchbase://couchbase:11210?detailed_errcodes=1');
+            $authenticator = new \Couchbase\PasswordAuthenticator();
+            $authenticator->username('Administrator')->password('password');
+            $cluster->authenticate($authenticator);
+            $bucket = $cluster->openBucket('default');
+            $collection = $bucket;
+            $bucketManager = $bucket->manager();
         } else {
             throw new Exception('ext-couchbase is not installed.');
         }
 
-        $authenticator = new \Couchbase\PasswordAuthenticator();
-        $authenticator->username('Administrator')->password('password');
+        if (!$this->waitForHealthyServer($bucket)) {
+            throw new Exception('Couchbase server is not healthy.');
+        }
 
-        $cluster->authenticate($authenticator);
-        $bucket = $cluster->openBucket('default');
-
-        $healthy = $this->waitForHealthyServer($bucket);
-
-        parent::__construct(new \MatthiasMullie\Scrapbook\Adapters\Couchbase($bucket, !$healthy));
+        parent::__construct(
+            new \MatthiasMullie\Scrapbook\Adapters\Couchbase(
+                $collection,
+                $bucketManager,
+                $bucket,
+                30000
+            )
+        );
     }
 
     /**
-     * Wait 10 seconds should nodes not be healthy; they may be warming up.
-     *
-     * @return bool
+     * Wait 30 seconds should nodes not be healthy; they may be warming up.
      */
-    protected function waitForHealthyServer(\CouchbaseBucket $bucket)
+    protected function waitForHealthyServer(/* \CouchbaseBucket|\Couchbase\Bucket */ $bucket)
     {
-        for ($i = 0; $i < 10; ++$i) {
+        for ($i = 0; $i < 30; ++$i) {
             $healthy = true;
-            $info = $bucket->manager()->info();
-            foreach ($info['nodes'] as $node) {
-                $healthy = $healthy && 'healthy' === $node['status'];
+
+            if ($bucket instanceof \CouchbaseBucket) {
+                $info = $bucket->manager()->info();
+                foreach ($info['nodes'] as $node) {
+                    $healthy = $healthy && 'healthy' === $node['status'];
+                }
+            } elseif ($bucket instanceof \Couchbase\Bucket) {
+                $info = $bucket->ping();
+                foreach ($info['services']['kv'] as $kv) {
+                    $healthy = $healthy && 'ok' === $kv['state'];
+                }
             }
 
             if ($healthy) {